combine offline and online locations to one component
