+++ packages/client/src/pages/location/location.tsx
@@ -3,8 +3,7 @@ import React, { lazy, Suspense } from 'react'
 import { Route, Routes } from 'react-router-dom'
 
 import { LoadingCircle } from '@etherealengine/client-core/src/components/LoadingCircle'
-
-const LocationPage = lazy(() => import('./[locationName]'))
+import LocationPage from '@etherealengine/client-core/src/world/Location'
 
 const LocationRoutes = () => {
   return (
